Add spec helper, tests, and factories
This commit adds a spec helper that sets up everything for testing. It
also adds spec tests for all of the api functionality currently included
in the application.  As well it adds some necessary files in the example
project that are used during testing. It also adds a spec command to the
rake file so that you can easily run all of the spec tests for the
project.